Move common parsing functions to parser.rs
